# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human-readable message
        type: string
    type: object
  internal_features_lists_transport_http.CreateListRequest:
    properties:
      author_user_id:
        example: 13
        type: integer
      name:
        example: Работа
        maxLength: 100
        minLength: 1
        type: string
    required:
    - author_user_id
    - name
    type: object
  internal_features_lists_transport_http.CreateListResponse:
    properties:
      author_user_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      version:
        type: integer
    type: object
  internal_features_lists_transport_http.GetListResponse:
    properties:
      author_user_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      version:
        type: integer
    type: object
  internal_features_lists_transport_http.ListDTOResponse:
    properties:
      author_user_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      version:
        type: integer
    type: object
  internal_features_lists_transport_http.PatchListRequest:
    properties:
      name:
        example: Работа
        type: string
    type: object
  internal_features_lists_transport_http.PatchListResponse:
    properties:
      author_user_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      version:
        type: integer
    type: object
...
  internal_features_tasks_transport_http.CreateTaskRequest:
    properties:
      author_user_id:
        example: 16
        type: integer
      description:
        example: Погулять сегодня вечером c собакой
        maxLength: 1000
        minLength: 1
        type: string
      due_date:
        example: "2026-04-03"
        type: string
      is_important:
        example: false
        type: boolean
      is_in_my_day:
        example: true
        type: boolean
      list_id:
        example: 14
        type: integer
      title:
        example: Погулять c собакой
        maxLength: 100
        minLength: 1
        type: string
    required:
    - author_user_id
    - is_important
    - is_in_my_day
    - title
    type: object
  internal_features_tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        type: integer
      complited:
        type: boolean
      complited_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      is_important:
        type: boolean
      is_in_my_day:
        type: boolean
      list_id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        type: integer
      complited:
        type: boolean
      complited_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      is_important:
        type: boolean
      is_in_my_day:
        type: boolean
      list_id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_tasks_transport_http.PatchTaskRequest:
    properties:
      complited:
        example: false
        type: boolean
      description:
        example: Одеть собаке шлейку и пойти c ней в парк
        type: string
      due_date:
        example: "2026-04-03"
        type: string
      is_important:
        example: true
        type: boolean
      is_in_my_day:
        example: false
        type: boolean
      list_id:
        example: 31
        type: integer
      title:
        example: Погулять c собакой
        type: string
    type: object
  internal_features_tasks_transport_http.PatchTaskResponse:
    properties:
      author_user_id:
        type: integer
      complited:
        type: boolean
      complited_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      is_important:
        type: boolean
      is_in_my_day:
        type: boolean
      list_id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_tasks_transport_http.TaskDTOResponse:
    properties:
      author_user_id:
        type: integer
      complited:
        type: boolean
      complited_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      is_important:
        type: boolean
      is_in_my_day:
        type: boolean
      list_id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
...
host: localhost:5050
info:
  contact: {}
  description: Todo Application REST-API scheme
  title: Golang Todo API
  version: "1.0"
paths:
  /lists:
    get:
      description: Просмотр списка списков))) с опциональной фильтрацией по автору
      parameters:
      - description: Получение списков конкретного пользователя по его id
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список списков))) успешно отдан
          schema:
            items:
              $ref: '#/definitions/internal_features_lists_transport_http.ListDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
      summary: Список списков)))
      tags:
      - lists
    post:
      consumes:
      - application/json
      description: Создать новый список в системе
      parameters:
      - description: CreateList тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_lists_transport_http.CreateListRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданный список
          schema:
            $ref: '#/definitions/internal_features_lists_transport_http.CreateListResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Author not found
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
      summary: Cоздать список
      tags:
      - lists
  /lists/{id}:
    delete:
      description: Удалить существующий в системе список по его ID
      parameters:
      - description: ID удаляемого списка
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление списка
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: List not found
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
      summary: Удалить список
      tags:
      - lists
    get:
      description: Получение конкретного списка по его ID
      parameters:
      - description: ID получаемого списка
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список успешно найден
          schema:
            $ref: '#/definitions/internal_features_lists_transport_http.GetListResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: List not found
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
      summary: Получить список
      tags:
      - lists
    patch:
      consumes:
      - application/json
      description: |-
        Измение информации об уже существующем в системе списке
        ### Логика обновления полей (Three-state logic):
        1. **Поле не передано**: `name` игнорируется, значение в БД не меняется
        2. **Явно передано значение**: `"name": "Работа"` - устанавливает новое имя списка в БД
        3. **Передан null**: `"name": null` - очищает поле в БД (set to NULL)
      parameters:
      - description: ID изменяемого списка
        in: path
        name: id
        required: true
        type: integer
      - description: PatchList тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_lists_transport_http.PatchListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно изменнённый список
          schema:
            $ref: '#/definitions/internal_features_lists_transport_http.PatchListResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: List not found
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Сonflict
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
      summary: Измененить список
      tags:
      - lists
...
  /tasks:
    get:
      description: Просмотр списка задач с опциональной пагинацией и/или фильтрацией
        по ID автора/списка
      parameters:
      - description: Фильтрация задач по ID автора
        in: query
        name: user_id
        type: integer
      - description: Фильтрация задач из по ID списка
        in: query
        name: list_id
        type: integer
      - description: Размер страницы с задачами
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с задачами
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список задач успешно отдан
          schema:
            items:
              $ref: '#/definitions/internal_features_tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
      summary: Список задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создать новую задачу в системе
      parameters:
      - description: CreateTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданная задача
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Author not found
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
      summary: Cоздать задачу
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Удалить существующую в системе задачу по её ID
      parameters:
      - description: ID удаляемой задачи
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление задачи
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
      summary: Удалить задачу
      tags:
      - tasks
    get:
      description: Получение конкретной задачи по её ID
      parameters:
      - description: ID получаемой задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Задача успешно найдена
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.GetTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
      summary: Получить задачу
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: |-
        Измение информации об уже существующей в системе задачи
        ### Логика обновления полей (Three-state logic):
        1. **Поле не передано**: `description` игнорируется, значение в БД не меняется
        2. **Явно передано значение**: `"description": "Одеть собаке шлейку и пойти с ней в парк"` - устанавливает новое описание в БД
        3. **Передан null**: `"description": null` - очищает поле в БД (set to NULL)
      parameters:
      - description: ID изменяемой задачи
        in: path
        name: id
        required: true
        type: integer
      - description: PatchTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно изменнённая задача
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Сonflict
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_med0viy_practika_internal_core_transport_http_response.ErrorResponse'
      summary: Измененить задачу
      tags:
      - tasks
...
swagger: "2.0"